Replace alfa review subcommands with flat flag-based command
