+++ packages/alfa-cli/src/alfa/command/review.ts
@@ -4,23 +4,16 @@ import { Option } from "@siteimprove/alfa-option";
 import { Arguments } from "./review/arguments.js";
 import { Flags } from "./review/flags.js";
 
-import answer from "./review/answer.js";
-import reset from "./review/reset.js";
-
 /**
  * @internal
  */
 export default (parent: Command) =>
-  Command.withSubcommandsAndArguments(
+  Command.withArguments(
     "review",
     parent.version,
     "Review potential accessibility issues by answering questions discovered during an audit.",
     Flags,
     Arguments,
-    (self) => ({
-      answer: answer(self),
-      reset: reset(self),
-    }),
     Option.of(parent),
     () =>
       async (...args) => {
