+++ packages/alfa-cli/src/alfa/command/review/arguments.ts
@@ -1,10 +1,1 @@
-import { Argument } from "@siteimprove/alfa-command";
-
-export const Arguments = {
-  url: Argument.string(
-    "url",
-    `The URL of the page to start an answer session for. Both remote and local
-    protocols are supported so the URL can either be an address of a remote
-    page or a path to a local file.`,
-  ),
-};
+export const Arguments = {};
