--- packages/alfa-cli/src/alfa/command/review/reset/arguments.ts
@@ -1,1 +1,0 @@
-export const Arguments = {};
