--- packages/alfa-cli/src/alfa/command/review/reset/flags.ts
@@ -1,5 +1,0 @@
-import { Flag } from "@siteimprove/alfa-command";
-
-export const Flags = {
-  help: Flag.help("Display the help information."),
-};
