+++ packages/alfa-cli/src/tsconfig.json
@@ -10,18 +10,10 @@
     "alfa/command/common/answering-oracle.ts",
     "alfa/command/common/question-store.ts",
     "alfa/command/common/recording-oracle.ts",
-    "alfa/command/review/answer/arguments.ts",
-    "alfa/command/review/answer/flags.ts",
-    "alfa/command/review/answer/run.ts",
-    "alfa/command/review/reset/arguments.ts",
-    "alfa/command/review/reset/flags.ts",
-    "alfa/command/review/reset/run.ts",
-    "alfa/command/review/answer.ts",
     "alfa/command/review/arguments.ts",
     "alfa/command/common/digest.ts",
     "alfa/command/review/flags.ts",
     "alfa/command/review/question-descriptions.ts",
-    "alfa/command/review/reset.ts",
     "alfa/command/review/run.ts",
     "alfa/command/review/utils.ts",
     "alfa/command/query/arguments.ts",
